staage set for c-dist with IR subtraction
git-svn-id: svn+ssh://fwe02.ifj.edu.pl/xdat1/lhcb/mchrzaszcz/svn/KK-all-2012@840 957d80b3-de0b-4737-8e94-343ebb428fb4

+++ MaMar/TMCgenFOAM.h
@@ -158,7 +158,7 @@ class TMCgenFOAM :public TMCgen
   void GetRhoFSR1(double, double, double&, double& );
   void GetRhoISR1(double, double, double&, double& );
   void GetRhoIFI1(double, double, double&, double& );
-  void GetRhoIFI1c(double, double, double& );
+  void GetRhoIFI1c(int, double, double& );
 
   double Soft_yfs(double );
   void MapPlus(  double, double, double &, double &);
